requirements.txt
pytest==7.3.1
selene==2.0.0b17
allure-pytest==2.13.2
pip==23.1.2
selenium~=4.9.1
requests~=2.31.0
python-dotenv~=1.0.0
voluptuous~=0.13.1
pytest-voluptuous==1.2.0